# Schemas/formComsent.cs
using server_truyen_v.Models;

namespace server_truyen_v.Schemas
{
    public class formComsent

    {

        public int id { get; set; } = 0;

        public string comment { get; set; } = "";

        public double ratting { get; set; } = 0.0f;

    }
}