# Schemas/formUpdateStory.cs
namespace server_truyen_v.Schemas
{

	public class formUpdateStory
	{
		public string Id { get; set; } = "";
		public string StoryName { get; set; } = "";
		public string CoverImage { get; set; } = "";
		public string Description { get; set; } = "";
		public string StoryAuthor { get; set; } = "";
		public string TypeDetailStory { get; set; } = "";

	}

}