# bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "API",
    "version": "v1"
  },
  "paths": {
    "/api/Comment/insert-comment": {
      "post": {
        "tags": [
          "Comment"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/formComsent"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Comment/update-comment": {
      "put": {
        "tags": [
          "Comment"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/formComsent"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Comment/delete-comment/{id}": {
      "delete": {
        "tags": [
          "Comment"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Comment/get-all-comment": {
      "get": {
        "tags": [
          "Comment"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/DetailStory/upload/{id}/{stt}": {
      "post": {
        "tags": [
          "DetailStory"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "description": "Nhập token dưới dạng 'Bearer {token}'",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "file",
            "description": "File upload",
            "type": "file"
          },
          {
            "in": "formData",
            "name": "id",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "stt",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/DetailStory/get-all-detailstory": {
      "get": {
        "tags": [
          "DetailStory"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/DetailStory/get-detailstory/{id}": {
      "get": {
        "tags": [
          "DetailStory"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/DetailStory/get-chapter/{id}": {
      "get": {
        "tags": [
          "DetailStory"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/DetailStory/delete-detailstory/{id}": {
      "delete": {
        "tags": [
          "DetailStory"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/DetailStory/update-view/{id}": {
      "put": {
        "tags": [
          "DetailStory"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Login/login": {
      "post": {
        "tags": [
          "Login"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/formLogin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/ServiceStoty/get-statistic/top-type-view": {
      "get": {
        "tags": [
          "ServiceStoty"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/ServiceStoty/get-statistic/top-view": {
      "get": {
        "tags": [
          "ServiceStoty"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/ServiceStoty/get-statistic/total-type": {
      "get": {
        "tags": [
          "ServiceStoty"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Story/inserts": {
      "post": {
        "tags": [
          "Story"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/formStory"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Story/getAll": {
      "get": {
        "tags": [
          "Story"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "string"
          },
          {
            "in": "query",
            "name": "shortChar",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Story/get-top-view/{limit}/{type}": {
      "get": {
        "tags": [
          "Story"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "limit",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "type",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Story/find-Story/{search}": {
      "get": {
        "tags": [
          "Story"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "search",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Story/filter-story/{search}": {
      "get": {
        "tags": [
          "Story"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "search",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Story/delete-story/{id}": {
      "delete": {
        "tags": [
          "Story"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Story/update-story": {
      "put": {
        "tags": [
          "Story"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/formUpdateStory"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Story/get-type": {
      "get": {
        "tags": [
          "Story"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "formComsent": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "comment": {
          "type": "string"
        },
        "ratting": {
          "format": "double",
          "type": "number"
        }
      }
    },
    "formLogin": {
      "type": "object",
      "properties": {
        "userName": {
          "type": "string"
        },
        "pass": {
          "type": "string"
        }
      }
    },
    "formStory": {
      "type": "object",
      "properties": {
        "storyName": {
          "type": "string"
        },
        "storyCode": {
          "type": "string"
        },
        "autho": {
          "type": "string"
        },
        "imgCover": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "typeStory": {
          "type": "string"
        },
        "typeDetailStory": {
          "type": "string"
        }
      }
    },
    "formUpdateStory": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "storyName": {
          "type": "string"
        },
        "coverImage": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "storyAuthor": {
          "type": "string"
        },
        "typeDetailStory": {
          "type": "string"
        }
      }
    }
  },
  "securityDefinitions": {
    "Bearer": { }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}